add update button

+++ FacultyView.java
@@ -34,17 +34,20 @@ public class FacultyView extends JPanel implements ActionListener {
    //global variables
    private String [] columnNames = {"Title", "Abstract", "Citation", "Author"};
    private Object [][] data;
+   private JButton jbReload;
    private JButton jbDelete;
+   private JPanel jpButtons;
    private JTable table;
    private Database db;
    private DLUser dl;
+   
 
 
    /**
    *Constructor
    */
    public FacultyView(DLUser dl) {
-
+   
       setLayout(new BorderLayout());
       
       //title at top
@@ -71,7 +74,7 @@ public class FacultyView extends JPanel implements ActionListener {
             
             table = new JTable(new CustomTableModel(columnNames, data));
             table.setRowHeight(50);
-      
+         
             JScrollPane jsp = new JScrollPane(table);
             jsp.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_ALWAYS);
             add(jsp, BorderLayout.CENTER);
@@ -87,11 +90,11 @@ public class FacultyView extends JPanel implements ActionListener {
             
             table = new JTable(new CustomTableModel(columnNames, data));
             table.setRowHeight(50);
-      
+         
             JScrollPane jsp = new JScrollPane(table);
             jsp.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_ALWAYS);
             add(jsp, BorderLayout.CENTER);
-
+         
          }
          db.close();   
       }
@@ -100,35 +103,98 @@ public class FacultyView extends JPanel implements ActionListener {
       }
       
       //delete button on south
+      jpButtons = new JPanel();
       jbDelete = new JButton("Delete");
-      add(jbDelete, BorderLayout.SOUTH);
+      jbReload = new JButton("Update");
+      jpButtons.add(jbReload);
+      jpButtons.add(jbDelete);
+      add(jpButtons, BorderLayout.SOUTH);
       jbDelete.addActionListener(this);
+      jbReload.addActionListener(this);
       
    } //end constructor
    
       
    public void actionPerformed (ActionEvent ae) {
-      //get selected row index
-      int rowIndex = table.getSelectedRow();
       
-      try {
-         db.connect();
+      Object choice = ae.getSource();
+      
+      if (choice == jbDelete) {
+      
+         //get selected row index
+         int rowIndex = table.getSelectedRow();
+         System.out.println(rowIndex);
          
-         //get the title from the array in that row
-         Object selected = data [rowIndex][0];
-         String selectedTitle = selected.toString();
+         try {
+            db.connect();
+            
+            //get the title from the array in that row
+            Object selected = data [rowIndex][0];
+            String selectedTitle = selected.toString();
+            
+            //query to delete row with that title
+            String sql = "DELETE FROM papers WHERE title = '" + selectedTitle + "'";
+            db.setData(sql);
          
-         //query to delete row with that title
-         String sql = "DELETE FROM papers WHERE title = '" + selectedTitle + "'";
-         db.setData(sql);
-
-         
-         db.close();
+            
+            db.close();
+         }
+         catch(DLException dle) {
+            System.out.println("DL EXCEption");
+         }
       }
-      catch(DLException dle) {
-         System.out.println("DL EXCEption");
-      }
+      else if (choice == jbReload) {
+         db = new Database();
+         try {
+            db.connect();
+         
+         /**
+          *This is such bad coding practice
+          *I didn't know how else to make it work 
+          *I am so sorry
+          */
+            if (dl.getUsername().equals("stevezilora")) {
+               String sql = "SELECT p.title, p.abstract, p.citation, CONCAT(f.fname, ' ', f.lname) FROM papers p " +
+                         "INNER JOIN authorship a ON a.paperId = p.id " +
+                         "INNER JOIN faculty f ON a.facultyId = f.id " +
+                         "WHERE f.lname = 'Zilora'";
+            
+               ArrayList<ArrayList<String>> myData = db.getData(sql, true);
+               data = db.convert(myData);
+            
+            
+               table = new JTable(new CustomTableModel(columnNames, data));
+               table.setRowHeight(50);
+            
+               JScrollPane jsp = new JScrollPane(table);
+               jsp.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_ALWAYS);
+               add(jsp, BorderLayout.CENTER);
+            }
+            else if (dl.getUsername().equals("danbogard")) {
+               String sql = "SELECT p.title, p.abstract, p.citation, CONCAT(f.fname, ' ', f.lname) FROM papers p " +
+                         "INNER JOIN authorship a ON a.paperId = p.id " +
+                         "INNER JOIN faculty f ON a.facultyId = f.id " +
+                         "WHERE f.lname = 'bogaard'";
+            
+               ArrayList<ArrayList<String>> myData = db.getData(sql, true);
+               data = db.convert(myData);
+            
+               table = new JTable(new CustomTableModel(columnNames, data));
+               table.setRowHeight(50);
+            
+               JScrollPane jsp = new JScrollPane(table);
+               jsp.setVerticalScrollBarPolicy(JScrollPane.VERTICAL_SCROLLBAR_ALWAYS);
+               add(jsp, BorderLayout.CENTER);
+            
+            }
+            db.close();   
+         }
+         catch(DLException dle) {
+            System.out.println("DL Exception");
+         }
 
+      }
+   
    }
 
 } //end class facultyView
